Playing with hash table sorting
Took an AGE to realise that [ordered] actually means "order of insert", and not by key, nor by value.
None of the posts I saw took the trouble to test with unordered data, therefore seeming to draw some false conclusions. Perhaps I am the one with the wrong end of the stick.